fixed PinpointsRef::clear bug
calling clear used to only remove markers from the map, while not removing the markers from the ref itself. A few tests were added to highlight it (bug was only apparent when testing PinpointsRef after a clear), and then the bug was fixed. A few minor changes were done to PinpointsDBCommunicationHandler, to handle null values in DB and use format strings in the url instead of string concatenation