Colored Pinpoints, with area
the marker icon was changed, and has a different color depending on the user who put it. A disk corresponding the validity of the guess (distance < threshold in meters) is now displayed on the map, for every Pinpoint. To abstract over all of this, A class Pinpoint was created which is a composite of Marker + disk, the interface of MarkerManager remains unchanged, PinpointRef remains unchanged, except for the remove method which now takes a Pinpoint as argument and not a Marker anymore. Also, CoordinatesConvertionsUtil was adapted to return a point (not creating an anonymous class everytime) when converting a Geopoint. It should be mentionned that ConvertionUtil should be changed to return cartesian coordinates when converting a geopoint for a more direct relation between the threshold of the disks and the threshold used by the game model