bug fix settings exception + sound control + theme fix
we can now use settings properly, the bad cast was removed and the code simplified. The achievements related to settings were removed as they were the cause of code lengthening and their logic was bound to what caused the bad cast. Sound control toggles were added to switch on & off music and/or sound effects. the themes were not being applied correctly. A solution was found, but requires a call to applyChosenTheme() in onCreate() in ALL activities